+++ AngularClient/src/main/account/login.component.ts
@@ -33,10 +33,9 @@ export class LoginComponent implements OnInit {
           this.notificationService.notification.next("You have been logged in!");
 
           // Get return url, reset loginReturnUrl and navigate
-          //TODO: Check this?
-          const returnUrl = this.authService.loginReturnUrl || "/app/home";
+          const returnUrl = this.authService.loginReturnUrl || "";
           this.authService.loginReturnUrl = "";
-          this.router.navigate([returnUrl]);
+          this.router.navigateByUrl(returnUrl);
         });
     }
   }