+++ AngularClient/src/main/core/components/core.component.ts
@@ -8,6 +8,8 @@ import { User, AuthService, NotificationService } from "backbone-client-core";
 import { Subscription } from "rxjs";
 import { mergeMap, map, filter } from "rxjs/operators";
 
+import { LoginComponent } from "../../account/login.component";
+
 @Component({
   selector: "core",
   templateUrl: "core.component.html",
@@ -110,6 +112,13 @@ export class CoreComponent implements OnDestroy, OnInit {
       filter(event => event instanceof NavigationEnd),
       map(() => this.activatedRoute),
       map(route => {
+
+        // Login return url: If the user is not logged in and not on Login page, then set "login return url"
+        if (!this.authService.currentUser.isAuthenticated()
+          && this.activatedRoute.firstChild.component !== LoginComponent) {
+            this.authService.loginReturnUrl = this.router.url;
+        }
+
         this.pathChecker(); // for home and mobile
         while (route.firstChild) { route = route.firstChild; }
         return route;